requirements.txt
argh>=0.24.1
google-cloud-storage==1.26.0
PyYAML>=3.10
black
Django<3.0
djangorestframework~=3.12.2
django-bulk-update
django-cors-headers
django-extensions
django-filter
django-rest-swagger
djoser
dj-database-url
drf-flex-fields
factory-boy
freezegun
gunicorn
psycopg2-binary
whitenoise
django-extensions
django-debug-toolbar~=3.2
djangorestframework-gis
haversine
Shapely
shortuuid
sentry-sdk
django-prometheus
celery==4.4.7
watchdog


requests~=2.25.1
pytz~=2021.1
pip~=21.0.1
setuptools~=53.0.0